Migrate tmpfs_prefix() to util.py & use in testing
