+++ packages/taucmdr/cf/software/tau_installation.py
@@ -1448,7 +1448,7 @@ class TauInstallation(Installation):
                 break
         else:
             raise InternalError(f"None of '{appfile_flags}' found in '{cmd}'")
-        tau_appfile = os.path.join(util.mkdtemp(), appfile+".tau")
+        tau_appfile = os.path.join(util.mkdtemp().name, appfile+".tau")
         LOGGER.debug("Rewriting '%s' as '%s'", appfile, tau_appfile)
         with open(tau_appfile, 'w') as fout, open(appfile) as fin:
             for lineno, line in enumerate(fin, 1):
